moved catalogs, mapreg, photeq
